MODULES: Simplify `network` tweaks

+++ modules/service.sh
@@ -144,21 +144,32 @@ if [ -f "/proc/sys/net/ipv4/tcp_available_congestion_control" ]; then
 
     for algo in bbr2 bbr cubic bic westwood newreno; do
         if echo "$congestion" | grep -qw "$algo"; then
+            chmod 644 /proc/sys/net/ipv4/tcp_congestion_control
             echo "$algo" > /proc/sys/net/ipv4/tcp_congestion_control
+            chmod 444 /proc/sys/net/ipv4/tcp_congestion_control
             break
         fi
     done
 
-    for param in tcp_ecn tcp_sack tcp_fastopen tcp_low_latency tcp_timestamps tcp_syncookies; do
+    for param in tcp_low_latency tcp_ecn tcp_fastopen tcp_sack tcp_timestamps; do
         case $param in
             tcp_fastopen) value=3 ;;
             tcp_ecn | tcp_sack | tcp_low_latency) value=1 ;;
-            tcp_timestamps | tcp_syncookies) value=0 ;;
+            tcp_timestamps) value=0 ;;
         esac
-        echo "$value" > "/proc/sys/net/ipv4/$param"
+        target="/proc/sys/net/ipv4/$param"
+        if [ -f "$target" ]; then
+            chmod 644 "$target"
+            echo "$value" > "$target"
+            chmod 444 "$target"
+        fi
     done
 
-    echo "0" > /sys/module/rmnet_data/parameters/rmnet_data_log_level
+    if [ -f /sys/module/rmnet_data/parameters/rmnet_data_log_level ]; then
+        chmod 644 /sys/module/rmnet_data/parameters/rmnet_data_log_level
+        echo "0" > /sys/module/rmnet_data/parameters/rmnet_data_log_level
+        chmod 444 /sys/module/rmnet_data/parameters/rmnet_data_log_level
+    fi
 fi
 
 for gpu in /sys/class/kgsl/kgsl-3d0; do
